requirements.txt
setuptools
smbus2
python3-memcached
pyyaml
pyserial
hbmqtt
redis
